// dotnet/openapi/model/loginResponseServiceResponse.ts
/**
 * JobPortal Admin Production
 *
 * Contact: 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LoginResponse } from './loginResponse';


export interface LoginResponseServiceResponse { 
    message?: string | null;
    statusCode?: number;
    data?: LoginResponse;
}